fix: separate code into small hooks
